Fixes double-clicking in PaperLights Default tool
There was an issue where if it was handled as single-click, there would never be opportunity for double click.  The solution is to delay executing single-click handler until after doubleClickTime has expired.  Also onMouseMove immediately fires the single-click handler (if it is there) since a double-click is nullified by mouse moving between clicks.